Use the type uint32 for storing the number of labels.

+++ python/boomer/boosting/cpp/example_wise_statistics.h
@@ -29,11 +29,11 @@ namespace boosting {
 
             std::shared_ptr<Lapack> lapackPtr_;
 
-            intp numPredictions_;
+            uint32 numPredictions_;
 
             const uint32* labelIndices_;
 
-            intp numLabels_;
+            uint32 numLabels_;
 
             const float64* gradients_;
 
@@ -94,8 +94,8 @@ namespace boosting {
              *                              search
              */
             DenseExampleWiseRefinementSearchImpl(std::shared_ptr<AbstractExampleWiseRuleEvaluation> ruleEvaluationPtr,
-                                                 std::shared_ptr<Lapack> lapackPtr, intp numPredictions,
-                                                 const uint32* labelIndices, intp numLabels, const float64* gradients,
+                                                 std::shared_ptr<Lapack> lapackPtr, uint32 numPredictions,
+                                                 const uint32* labelIndices, uint32 numLabels, const float64* gradients,
                                                  const float64* totalSumsOfGradients, const float64* hessians,
                                                  const float64* totalSumsOfHessians);
 
@@ -196,7 +196,7 @@ namespace boosting {
 
             void updateCoveredStatistic(intp statisticIndex, uint32 weight, bool remove) override;
 
-            AbstractRefinementSearch* beginSearch(intp numLabelIndices, const uint32* labelIndices) override;
+            AbstractRefinementSearch* beginSearch(uint32 numLabelIndices, const uint32* labelIndices) override;
 
             void applyPrediction(intp statisticIndex, Prediction* prediction) override;
